move ops module to its own file
